Fix: Auth CRUD and Blog Crud

+++ app/controllers/blogs_controller.rb
@@ -25,10 +25,10 @@ class BlogsController < ApplicationController
   end
 
   def create
-    return if suspended(@current_user.can_blog_date)
+    # return if suspended(@current_user.can_post_date)
 
-    # @blog = Blog.new(blog_params.merge(user_id: @current_user.id))
-    @blog = @current_user.blogs.build(blog_params)
+    @blog = Blog.new(blog_params.merge(user_id: @current_user.id))
+    # @blog = @current_user.blog.build(blog_params)
     
     # if authorized?
       if @blog.save
@@ -56,19 +56,19 @@ class BlogsController < ApplicationController
     end
   end
 
-  def destroy
-    # Only allow the owner of the post or an administrator to destroy the post
-    unless @blog.user == @current_user || @current_user.admin_level >= 1
-      render json:{errors: 'Not authorized to perform this task'}, status:401
-    end
+  def destroy 
 
-    if @blog.destroy
-      render status: :ok, 
-        json: {notice:'Blog deleted'}
+    if authorized?
+      if @blog.destroy
+        render status: :ok, 
+          json: {notice:'Blog deleted'}
+      else
+        render status: :unprocessable_entity,
+          json: {errors: @blog.errors.full_messages.to_sentence}
+      end 
     else
-      render status: :unprocessable_entity,
-        json: {errors: @blog.errors.full_messages.to_sentence}
-    end  
+      handle_unauthorized
+    end    
   end
 
   def pin_blog
@@ -111,15 +111,17 @@ class BlogsController < ApplicationController
     false
   end
 
-  # def authorized?
-  #   #  @blog.user == @current_user
-  #    @blog.user_id == @current_user.id
-  # end
+  def authorized?
+    # Only allow the owner of the post or an administrator to destroy the post
 
-  # def handle_unauthorized
-  #   unless authorized?
-  #     render json:{notice:"Not authorized to perform this task"}, status:401
-  #   end
-  # end
+    #  @blog.user == @current_user ||  @current_user.admin_level >= 1
+     @blog.user_id == @current_user.id || @current_user.admin_level >= 1
+  end
+
+  def handle_unauthorized
+    unless authorized?
+      render json:{notice:"Not authorized to perform this task"}, status:401
+    end
+  end
   
 end
